SampleDataTest: Move to systemtests
As AddressBookGuiTest and AddressBookSystemTest are very similar, it’s
not necessary to have both of them.

As part of removing AddressBookGuiTest, let’s move SampleDataTest to
systemtests.
