+++ src/test/java/systemtests/SampleDataTest.java
@@ -1,4 +1,4 @@
-package guitests;
+package systemtests;
 
 import static seedu.address.ui.testutil.GuiTestAssert.assertListMatching;
 
@@ -13,16 +13,20 @@ import seedu.address.model.person.Person;
 import seedu.address.model.util.SampleDataUtil;
 import seedu.address.testutil.TestUtil;
 
-public class SampleDataTest extends AddressBookGuiTest {
+public class SampleDataTest extends AddressBookSystemTest {
+    /**
+     * Returns null to force test app to load data of the file in {@code getDataFileLocation()}.
+     */
     @Override
     protected AddressBook getInitialData() {
-        // return null to force test app to load data from file only
         return null;
     }
 
+    /**
+     * Returns a non-existent file location to force test app to load sample data.
+     */
     @Override
     protected String getDataFileLocation() {
-        // return a non-existent file location to force test app to load sample data
         String filePath = TestUtil.getFilePathInSandboxFolder("SomeFileThatDoesNotExist1234567890.xml");
         deleteFileIfExists(filePath);
         return filePath;
